feat: protect delete room endpoint

+++ src/pages/api/sessions/[sessionId]/rooms/[roomId].ts
@@ -1,10 +1,14 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
+import { getSession } from 'next-auth/react'
 import addRoom from 'src/api/rooms/addRoom'
 import deleteRoom from 'src/api/rooms/deleteRoom'
 import getRoom from 'src/api/rooms/getRoom'
 import updateRoom from 'src/api/rooms/updateRoom'
+import { UserRoles } from 'src/utils/enums/userRoles'
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
+  const session = await getSession({ req })
+
   switch (req.method) {
     case 'GET': {
       return getRoom(req, res)
@@ -19,6 +23,13 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
     }
 
     case 'DELETE': {
+      if (session?.user.role !== UserRoles.ADMIN) {
+        return res.status(401).json({
+          message: 'You are not authorized to delete rooms',
+          success: false,
+        })
+      }
+
       return deleteRoom(req, res)
     }
   }
